chore: move ProcessorConfiguration out of internal package
